pyproject.toml
[tool.poetry]
name = "django-phonebook"
version = "0.1.0"
description = ""
authors = ["Drixler Angelo B. Historillo <drixlerangelo.historillo@ingrammicro.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "~5"
djangorestframework = "~3"
psycopg2 = "~2"
python-dotenv = "^1.0.1"
django-oauth-toolkit = "^2.3.0"
django-allauth = {extras = ["mfa", "socialaccount"], version = "^0.63.0"}
drf-spectacular = "^0.27.2"
django-filter = "^24.2"
pandas = "^2.2.2"
openpyxl = "^3.1.4"
pusher = "^3.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
